MusicXML: Fix chord conversion
Fix the conversion of chords from MusicXML to lilypond. The problem was
that the #text entries in the XML tree (converted to a musicxml.Hash_text
python instance) messed up the tracking of the current positions
in Part.interpret. As soon as these children are ignored, the original
algorithm to assign measure positions works again.

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>