MusicXML: A grace note at the beginning shall not set staff for the voice
The sample files from Recordare contain a piece where the voice on the
upper staff of a PianoStaff starts with a grace displayed in the second
staff. So far, this put the whole voice on the second staff. Thus, grace
notes are now ignored when determining the voice for a staff.

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>