MusicXML: Fix lyrics on chords, ignore lyrics for notes inside melismata
The lyrics detection code so far ignored chords (ie. it added a \skip for
every additional note in the chord). Now I simply ignore all subsequent
notes and the corresponding syllables for the lyrics.

By default, lilypond does not add lyrics to notes in melismata (i.e.
tied notes and notes inside slurs), so while walking through all notes,
we need to check whether we are inside a slur, a tie or a grace note
and simply ignore the lyrics on these notes.

There are some scores, which still put lyrics on notes inside slurs, so
in the future, I'll also add a way to set/unset ignoreMelismata, but
that's left for the future.