MusicXML: Fix small problems with assignment <harmony> <=> voice
Oops, I didn't want to treat <direction > elements like <harmony>.
The direction elements should still be asssigned to all voices (or
all voices of the staff).

Also, at the end of the loop, assign all remaining <harmony> elements
to the voice of the previous note and don't simply discard it.