MusicXML: Add possibility for additional defs, implement snap-pizzicato
Lilypond does not have all definitions that we might need for the score,
so I now also keep a list of additional definitions and insert the
corresponding lilypond code at the beginning of the *-defs.ly file.

One example is the snap-pizzicato markup (see Snippet 335 on LSR),
which is also implemented with this commit.