MusicXML: Implement relative mode for pitches
Relative mode for pitches is now triggered by an -r or --relative
command-line switch.

In this case, each voice is wrapped inside a RelativeMusic wrapper
object, and the previously converted pitch is stored in a global state
variable. The case of chords needs to be handled with special care,
as the whole chord uses the pitch of the first note as its base pitch.
So, after converting all notes in the chord, we have to reset the
global state variable to the pitch of the first note.

Also added myselfs to the copyright output.