MusicXML: Fix using lxml as XML parser
Unfortunately, the default lxml.etree parser also includes comments
in its tree, which does not return a string as its tag, but a built-in
function Comment. Thus, I need to filter out all these comments.
It seems that lxml does not have any convenient way to check whether
a given node is a comment, so I convert it to its string representation
and compare with "<!--.*-->". Of course, this is a bad hack, but I haven't
found a better way yet.