Show TupletNumbers also on single-note tuplets (e.g. with tremolo)
While tuplet brackets should not be printed for tuplets consisting
of only one note (e.g. a note with tremolo!), the tuplet number still
needs to be printed and can not be ignored. Otherwise the timing would
look messed up.

Since the tuplet number also uses the dimensions calculated by the tuplet
bracket, the tuplet bracket at least needs to calculate it's theoretical
size, too, even though it will not be printed for single-note tuplets. The
check to kill zero-length tuplet brackets has thus been moved from
calc_positions to print in Tuplet_bracket.

In the print function we mustn't kill the bracket (which
will also kill the number), but only hide the bracket.

Also added a test case.