lilypond-book: store document language in a global variable
Document language as 'lang' ly option caused trouble, especially
different checksums for identical snippets depending on the language
of the Texinfo document.

Using a global variable is not ideal either, but as processed documents
are not implemented as classes, there is currently no other way to store
that kind of info and make sure that included files inherit it.