texi2html: Logic to load translations (secname -> filename map)
In the future, section names will be translated directly in the
.tely file, with a @translationof command indicating the original
section name (which will be used for the file name). Since we are
also putting several sections into the same file, the file name
can no longer be obtained from the section name only. Instead,
we need to keep a map of secname->filename#anchor.

This commit adds the general structure of the logic to look up the
correct file name and anchor for sections, although loading
and reassigning the new values is not yet implemented.