Add two-parameter versions of @ruser etc. to allow specifying the displayed text
Sometimes node named cannot contain special characters, which you would like
to be displayed in the link text. Thus, the third argument of @ref needs to be
used. Unfortunately, Texinfo (or rather texi2pdf) does not support optional
arguments to macros, so we have to add new two-parameter versions of all our
xref macros.

I als streamlined the definitions in macros.itely a little bit (distinguish only bigpage, not bigpage and tex).

This also allows us to link to the Top node of any manual with displaying the
manual title instead of "Top".