Add a make-line-stencil function, which correctly sets stencil extents
So far, one had to create line stencils manually using
   (ly:make-stencil (list 'draw-line ...) xext yext)
Unfortunately that meant that one had to specify the x- and y- coordinates
twice. This new make-line-stencil function takes the coordinates once,
creates the stencil and properly sets its extent (adding half the line
width to all coordinates of the x- and y-intervals).

Signed-off-by: Reinhold Kainhofer <reinhold@kainhofer.com>