new showFirstLength feature
This commit implements a showFirstLength property, that does
the opposite of showLastLength (both properties may be set,
in this case the beginning and the end of the piece are printed,
with a double barline as a separator).

When only showFirstLength is set, the 'length property is
overriden (except if already shorter than the specified first
length) and therefore the compilation will be much faster.

This new feature may be very convenient, e.g. to print the
index of a book containing only the beginning of each \score.