Fix poor automatic fretboard calculation
Improve the automatic fretboard calculation algorithm by changing
the sort order of the notes.

This improvement broke the example on improving automatic fretboards
by specifying a string.  I could not find another example, so I
removed the existing example.  I wanted to leave the text in case
somebody had the problem in the future.

This improvement also fixed the poor diagram for c:9 in the
predefined fretboards section of fretted-strings.itely