MusicXML: Fix clefs that appear in only one staff (ignore for the other!)
This bug caused a position-off problem, which messed up chord detection
and caused several other problems. If a clef has a number attribute
(indicating its staff number), only copy it to the relevant staff.
For all other staves, if the remaining attributes are empty, simply
ignore them instead of passing an emtpy <attributes> tag, which was
the real cause for the problem.