lys-to-tely: If a .i?te(ly|xi) file is passed to the script, include it
lsy-to-tely simply creates one texinfo page from a list of snippets. If
one wants to insert some texinfo code in between some snippets,
that's currently not really possible (only as texidoc header field of
some dummy snippet, that will create an empty image and display the
dummy file name). This patch allows *.i?te(ly|xi) files to be passed
to lys-to-tely, too, which will simply be include via
@include ${filename}
in the .tely file. All other files are still treated as lilypond snippets
with no difference to the current state.

lys-to-tely will use the same order of files as passed on the command line,
so by sorting the file names appropriately, one can now insert
arbitrary texinfo code somewhere in between some snippets (or at beginning
or end of the file).