lilypond-book: Pass lilypond version to lilypond-book via env var in makefiles
When lilypond-book.py is called from our makefiles to build the documentation,
the @TOPLEVEL_VERSION@ string has not been expanded and the lilypond
version will not be available to lilypond-book.

This patch detects this case and uses the LILYPOND_VERSION environment
variable in that case, which is passed to lilypond-book.py by the makefiles.