Fix #713
In Paper_book::output_aux(), for a parent Paper_book, despite the
looping:
  for (SCM p = scm_reverse (bookparts_); scm_is_pair (p); p = scm_cdr (p))
not all children bookparts are processed, causing the bus error to
occur later on.

As a workaround, bookparts will be stored in Paper_book in straight
order (in Book::process_bookparts), instead of reverse order, to avoid
the multiple reversings (in paper_book.cc).