Fix vertical alignment of FretBoards
The origin of the fret diagram will be at fret zero, string 1.
The x alignment is adjusted to put the fret diagram over a note
for the fret diagram markup.  The y alignment used to be centered,
but now it is not adjusted during creation of the grob.

Add regression test for fret board alignment