Fix error message in define-markup-function
There was a syntax error in markup.scm
Also, ly_error would not pass the arguments to simple-format properly
if rest was a list in the scheme side (as in define-markup-function).
Therefore, if the format args are a list, get the car of rest in
ly_error so the args can be passed properly.