Move left-broken line-spanner check to callback.
Currently, left-broken line spanners and hairpins are removed when they
end on the first note of the new system.  Though this is appropriate
for glissandi and voice followers, there is no way to override this
default behaviour.

This patch allows the user to tweak the appearance by overriding
'after-line-breaking.

- removed left-broken check from ly:line-spanner::print to a new
after-line-breaking callback, ly:spanner::kill-zero-spanned-time

- removed hairpin after-line-breaking callback and associated method
consider_suicide ()

- added new callback to Hairpin, Glissando, TrillSpanner and VoiceFollower

- added convert rule for ly:hairpin::after-line-breaking