Add lists for possible font encodings
Currently, there is no central location for all of the font encoding
values LilyPond uses.

This patch creates two new procedures, one that returns a list of all
encodings used for LilyPond's music fonts (Emmentaler and Aybabtu),
and one that returns a list of the encodings used for all other fonts
(currently, latin1).