SVG backend: Output a single SVG file for each page
If the SVG backend is used when creating multiple-page scores, a single
SVG file is created, using the <pageSet> and <page> elements to delimit
individual pages.

Unfortunately, these elements are not supported by any known user
agent, and they are part of a W3C Working Draft (SVG Print 1.2) that
might not reach recommendation status for several years.

At this point in time, a better solution is to output a single SVG file
for each individual page.  Outputters are created and closed for every
page, and the unique filename suffixes are used for each successive page.

We want to *eventually* add support for <pageSet> and <page> with a -d
option, so the source code comments have been updated to reflect this.