Fix handling of missing stencil expressions
I had not tested the case of multiple \book blocks.

For example, in output-ps.scm a warning is assigned for "utf-8-string",
but this assignment persists for the next \book block, which is
undesirable; the assignments are only supposed to be used while the
stencils are being dumped with the outputter.

This patch removes the "missing" stencil expressions from the
output-module when the outputter is closed.