Simplify the handling of neighbor columns.
Now we keep track of only one left-neighbor and one right-neighbor
for each column.  This makes the loose column determination simpler
and less fragile.  Also, we now use a spacing wish if it has
any right-item in the appropriate column (before, we only used it
if its left-most right-item was in the appropriate column).  This
fixes the ad-hoc loose-column determination for strict note
spacing.