Web and doc build: fix IE rendering
If *any* markup is found before the DOCTYPE tag, all versions of IE will
fall into a type of severe "quirks" mode.  It's not pretty.

Thus, we will only add <!-- header_tag --> to pages without DOCTYPE tags
(primarily those not generated by texi2html).

This fixes IE 7 and 8 for the new website and the documentation.