Improve behavior of `make distclean'
This target is supposed to perform a `make clean', and then remove the
appropriate parts of configure output.

Thus, $(LOOP) should be removed, since configure output only appears in
the top-src-dir, and $(outdir) should be removed afterwards.

The other two changes are just nitpicks: the quotes are not needed, and
$(outdir) already expands to ./out, so the extra ./ is superfluous.