Vim ftdetect: use `set ft' instead of `setf'.
Since filetype.vim is no longer distributed, the `setf' command will
only set the filetype if Vim's default filetype.vim does not detect a
matching filetype first.

Unfortunately, when certain lines begin with #(, the "generic" filetype
is matched, so the LilyPond-specific highlighting is never loaded.

A workaround is to use `set ft'; this sets the filetype to `lilypond'
unconditionally for all files ending in .ly or .ily, which is what we
want.