Fix check for `utf-8-string' in output modules
When checking for `utf-8-string', we must not include the "warning"
definitions (set at runtime) for backends without `utf-8-string'.

An additional check is therefore necessary in order to ignore module
variables bound to lambda expressions.

All of the stencil expression procedures defined in output-*.scm create
variable<->symbol bindings, and has_utf8_string should be true in these
cases.  But the variable<->#f bindings should not count, because they
are set at runtime.

This commit adds an additional check to disclude the variable<->#f
bindings.