Improve 'warning-as-error handling again
Currently, on the Scheme side, ly:warning and ly:programming-error are
redefined to ly:error.  This works, but if the user wants to override
this setting in an LY file, he will not be able to.

This commit moves the handling for the callbacks to C++, thus permitting
users to override -dwarning-as-error in LY files for calls to ly:warning
and ly:programming-error.