lilypond-book: Change md5 hashing strategy
Previously, the md5 hash for folder/file names was based on the entire
.ly file contents.  This worked okay until the preamble changed, which
then broke `make check'.

This commit changes the strategy so that only the LY code (without the
preamble) is used when computing the md5 hash.  This will fix the regtest
comparison output between releases from now on.