Docs: fix compile
One or both of these macros was breaking the build for the German NR
PDF.  There are other places in the docs with @rlearning commented out
for similar reasons, I think.

It seems that the custom @lilyvindex macro sometimes fails when its
argument begins with a non-ASCII character.  This is just a guess.