New `vspace' markup command, after Nicolas Sceaux
This commits adds a new builtin-markup-command,
as used by Nicolas in his scores.  This command
takes a positive number as argument, and makes
an invisible stencil of three times this number
(which is somehow equivalent to the vertical space
taken by a character).

The command definition includes a docstring and
an example, and changes.tely has been updated
accordingly.