Issue #768: Chord repetition shortcut
In the lexer, add a chord repetition state, holding the repetition
symbol (like note names, but for chord repetition), the repetition
function used to copy the previous chord, and a slot to save the
previous chord.  Use the repetition symbol to detect a chord
repetition when lexing.

In the parser, when a note chord is found, save it (it becomes
previous chord).  When a chord repetition token is found, make a chord
using the repetition function and the previous chord.

The repetition symbol and function are initialized in
ly/chord-repetition-init.ly. The repetition symbol defaults to `q' (as
quote or qord -- as suggested Werner). The function copies the
pitches, but not articulations.