Fixes issue 786, "Extenders in lyrics stop prematurely if a single underscore is found."
This commit changes the lyrics engraver to create LyricText objects
even for empty lyrics (underscores). This change is necessary because
the old behavior (pre-7531ea6b3 commit) relied on extenders continuing
until the presence of another lyric syllable, which was not always the case.
That commit changed the behavior to completize extenders when no more lyrics
were present, which fixed the neverending-extender bug but introduced the
extenders-stopping-prematurely bug.

By adding the empty LyricText objects, the extender engraver can now tell the
difference between melismas and the end of a block of lyrics.