Fix warning message in predefined fretboards
When a predefined fretboard is found, don't try to calculate
a fretboard.  Previously, even though the predefined fretboard
was used, the strings and frets for an automatic fretboard were
calculated.  This created a warning message if the chord notes used
were below the range of the instrument, even if a predefined fretboard
exists for those notes.