Fix #909.
In the acknowledger for the bar-line-interface, a check to see if the
acknowledged grob contains the bar-line-interface is superfluous.

Instead, filter out SpanBar grobs (with a span-bar-interface check) in
the acknowledger so that SpanBars are not added to bars_ if more than
one Span_bar_engraver exists in a context.