Web build: complicate macro to create links.
This ensures a more unified approach, but more importantly means
that any mistakes will be more visible.

Also, it paves the way to do a complicated @iftex, @ifhtml @ifset
web, etc, to change the @url links.  (I mean, the horrible
construction that's upcoming will only be horrible in one place,
instead of duplicating this 10 times in manusl.itexi and
community.itexi)