Partial revert of 7b34eac90f81c8123feaf2670fafae236101f14c.
Look folks, don't screw with the texinfo macros unless you really,
really test it before pushing.  These things are stupidly finicky
at the best of times.  Also, take it step by step.  Don't change a
whole bunch of things at once, because when (not if) it breaks
it's much harder to figure out how to fix it.