Web: remove comment-based bug reports.
Guys, don't use comments to discuss problems in the website.
Email lilypond-devel or add an issue in the tracker.  We spent a
lot of time trying to stop this practice during GDP.

I don't care if you want to have discussions via source code
comments for programming issues (although I it's a stupid way to
communicate), but keep this garbage out of the texinfo files.