Fix DynamicTextSpanner left alignment.
When a DynamicTextSpanner directly follows a DynamicScript, padding is
required to ensure the 'left 'text doesn't collide with the
dynamic sign.  Unfortunately, this default value skews the
spanner's left bound to the right when no dynamic script is present.

This patch caters for both situations by resetting 'left 'padding
and 'left 'attach-dir as required, adding extra padding read from the
property 'right-padding to fine-tune the space between the text elements.

* add regtests demonstrating 'right-padding and correct alignment

* remove hard-coded 'bound-details setting in New_dynamic_engraver, which
  prevented user override

* add dynamic-text-interface with user property 'right-padding

* add 'before-line-breaking callback dynamic-text-spanner::before-line-breaking
  which checks whether the left bound is a DynamicText; if this is true,
  set nested properties 'left 'attach-dir/'padding to prevent collision