Doc: new chapter "External programs" in Usage.
This consists of the old "Converting from other formats" chapter,
plus the "Point and click" and "Text editor support" from the
Running lilypond chapter; all of these kind-of stuck out
awkwardly.

No nodes were harmed (renamed) in the making of this patch so far.

"Running lilypond-book"'s "Alternate methods..." has been moved
into External programs, with the new name of "LilyPond output in
other programs".  The old name still exists for ease of TOC and
whatnot, but anybody referring to it might want to change to the
new name to avoid a relatively pointless redirect.