Build: another ugly hack for website integration.
My comment for issue 1004 (opened for this patch):
As release manager, I feel that I should reject the patches I'm
writing because they are, as "weegies" (people from Glasgow) would
put it, "utter pish".  However, as overall manager, I want to let
doc writers and translators work on the website and see how it
really looks.

Given my doc roots, the doc side wins, so I'm hacking stuff.  But
only with the promise that this item is High priority, and that
I'll work on it as soon as I've finished all Critical items.  It
*will* be fixed before GLISS starts.  GOP might start first, if
everything else is already in place.