Instanciable scheme engravers
A scheme engraver definition may be a one argument procedure,
taking the context where it is defined as an argument, and
evaluating to an a-list scheme engraver definition.
This allows to define instance slots for a scheme engraver.

This patch also defines some scheme callbacks for grob related
functions, and add an example snippet showing how an existing
C++ engraver may be defined in Scheme: the ambitus engraver.