Fix #1020.
The default size of the Century Schoolbook fonts is 11pt (as seen in
font.scm).

But when a 'font-name is specified, the 'text-font-size property is
used, which changes the default size to 12pt.

This commit standardizes the default size to 11pt, and also fixes the
behavior of \abs-fontsize when setting font size to 12pt.

Note: there are still some rounding errors related to converting between
LilyPond's scale and Pango's scale, but they are relatively small.