Add lilypond-book ly verbatim gettext quirk
Documentation/snippets/jazz-combo-template.ly contains a variable
definition named "bass", of which lilypond-book translates all
occurences in Texinfo verbatim ly code, including the case where
"bass" is an argument of the \clef command.  As variables may be used
in Scheme code and this gettextization is done without real parsing,
it's impossible to avoid such corner cases, which will thus be fixed
ad-hoc as they are discovered.

Thanks to Francisco Vila for the report.