Doc: revert bad fix for 617.
Neil said "This is not recommended at all.  The default callback
is a closure, combining the result of
ly:self-alignment-interface:x-aligned-on-self *and*
ly:break-alignable-interface::self-align-callback.  Thus
overriding X-offset also causes 'break-align-symbols to be
ignored, messing up the positioning completely."

I don't understand it, but I understand that he knows more than
me.  :)    I've left the little clean-up that James did.