Lilypond-book: Implement MusicXML support in lilypond-book
This patch adds support for including MusicXML files into
documents processed by lilypond-book. In particular:
-) HTML: <musicxmlfile options>filename.xml</musicxmlfile>
-) TeX: \musicxmlfile[options]{filename.xml}
-) Texinfo: @musicxmlfile[options]{filename.xml}

Since MusicXML is so verbose, it doesn't make much sense to
support inline MusicXML.

The snippets are basically processed like a lilypond file,
except that musicxml2ly is run on them (with the options given
for the snippet) and the returned lilypond code is then processed
as if it were the original contents of the file.

For the output links, however, the html and texinfo pages will
link to the .xml/.mxl file rather than the intermediate .ly file.

If a file has the extension .mxl, it is assumed to be a compressed
MusicXML file (alternatively, the 'compressed' snippet option can
be given).

What's missing is proper documentation in "Usage". I'm unsure
how to document such new snippet types...