Fix #413.
The `convert-to-png' procedure did not account for the 'landscape
setting, so a portrait orientation was always used.

This commit fixes the issue, so that width and height dimensions are
swapped whenever landscape orientation is set for full-page PNG output.

Additionally, I extracted some common routines from `convert-to-pdf' and
`convert-to-png' into three separate procedures: `output-width-height',
`output-resolution', and `output-filename'.