Add include-settings cmd line option to include a custom settings
This allows the user to apply the same set of options
to different files, or to apply different settings to
the same score without modifying the score itself.
In particular,
  lilypond -dinclude-settings=myinclude.ily file.ly
will compile file.ly, but read and execute myinclude.ily
before file.ly is processed. That way, e.g. global paper
changes can be easily done, and a file can be run several
times, each time with a different include-settings file
containing different settings.

This feature is similar to the --init=DATEI feature,
except that --init will replace the whole init sequence
(thus your init file will need to copy all of lilypond's
init.ly), while -dinclude-settings will only add some
definitions to the normal init sequence in init.ly (i.e.
the file is included in init.ly, so the normal init sequence
takes place).

Unfortunately, it is not possible to add lilypond variable
definitions for use in the score... This is probably due
to the way the init sequence works.