Lilypond-book: Fix problem with os.path.abspath in Windows
The check whether input file == output file needs to use the absolute
path to the input file, since the CWD has changed meanwhile (and is
the output dir at that point). Unfortunately, we can't call os.path.abspath
right at the beginning and use the abspath for all of the processing,
because os.path.abspath will replace forward slashes by backslashes
in Windows (the path is normalized!). That path with backslashes
are then printed to the snippet map files and cause l-p to fail with
wrongly escaped characters...

As a solution, use the relative pathname of the input file for all
processing, but also store the abspath for that one particular
infile==outfile check.