LSR: escape double backslashes.
4+4+10 backslashes on one line to do this?!  How the bloody mao
can people (including myself) still refer to python as a "highly
readable language" ?

Also, I'm appalled at the state of makelsr.py.  A few years ago it
was a nice and simple script; now it's a confusing mess of regex,
some stuff in functions, some stuff in global, etc.  It took me
over an hour to add the "escape double backslahes" to this script.

Granted, I'm an idiot without previous experience with regex...
but hey, I'm the only person to work on this bug in the over 3
years.  I think our python+build scripts need to be written with
idiots in mind, because smart people are too busy doing other
stuff.