Improve pure-height estimations with outside-staff-priority.
Since outside-staff placing using outside-staff-priority can
only be done properly when horizontal positioning is available,
the pure-height estimates did not take outside-staff-priority
into account at all.  This commit does a better approximation
to outside-staff placement, in which the outside-staff placements
are computed assuming that the outside-staff grobs in one bar
do not collide with those in the next bar.

The estimates are still bad for RehearsalMark and other grobs
that live at the Score level.