convert-ly: option to only modify changed files.
This useful for the regtests -- with the extra -d flag, we will
only change the \version string if the actual contents of the file
has changed.  This lets us see which tests were modified, and
which tests were unchanged since the previous stable release.

However, the version string will still be updated to match the
last stable release.