Send ping logs to /dev/stderr instead of /dev/tty.
(This fixes issue #1028)

If `-dlog-file' was not specified along with `-dseparate-log-files',
certain log messages (the "ping" log messages) were sent to /dev/tty.

However, this was causing problems for machines building LilyPond that
could not write to /dev/tty (no controlling tty).

The solution is to write to /dev/stderr instead, which is
common-practice for logging.

See the Debian bug report for details:

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=572290