Indent lily-git.tcl
This is the only tcl file in our repo.  It doesn't have a long
history, nor is this file at all important to compiling+debugging
lilypond.  As such, I feel justified in tossing it through emacs'
indentation.

This does not mean that I think it is appropriate to discuss the
indentation of any other type of file in git.  If anybody
complaints about indenting this file, I will remove this commit
without question, and we will keep on using the previous
inconsistently-indented version until we discuss code indentation
as part of GOP.