manual convert-ly: minimum-Y-extent (1298+1299).
Patch by Keith O'Hara.

cary.ly used Y-extent to increase space for staves, use
markup-system-spacing instead.

orchestra.ly used Y-extent to created a custom context for rows
rehearsal and tempo marks.  The new way to creat such a row is to
give the context a staff-affinity, and put the spacing
requirements into nonstaff-unrelatedstaff-spacing, etc.  This file
also had its own definition of the \Dynamics context, which is now
built-in.

percussion.itely used Y-extent to produce a spacing to the lyrics
line very similar to what we now get by default.

spacing.itely used Y-extent among other \overrides in an example

ancient.itely used Y-extent to make an invisible staff take up no
space.  The markup can stand alone with no staff.  (If for some
reason a @lilypond must contain notes or skips, then just delete
the 2 lines with minimum-Y-extent.)

The relevant portion of gregorian.ly defined a layout for printing
naked neumes, with no visible staff.  The function of that line,
setting the height of that invisible staff to zero, is not
actually used anywhere in the Google-able portion of lilypond.org,
but the 2.13 equivalent gives the same output the old method, in
case anyone uses this layout to make a vertical stack of naked
neumes.