FiguredBass: Extenders for figs of different width should stop at same position
FB extenders use the extent of the hidden figure grob, so extenders
of different figures might stop at different positions. Ideally, the
note of the associated staff should be retrieved and used as the
position to stop the extender. However, FB can also exist standalone
without any attached staff.

So, as a workaround, this patch simply sets the text of all hidden
figures to 0, so they all have the same extent and all extenders will
stop at the same position.