Ties: Print out a warning for unterminated ties
Do not silently clear a tie attempt, but notify the user that a tie
could not be created (typically either due to a typo or due to a missing
tieWaitForNote)

In a chord, a tie is only unterminated if none of the notes of the
chord create a tie. All other notes of that chord are possible
candidates for a tie start, but not neccesary. For this, I simply
suppress the warning if a note from the same moment already caused
a tie.
E.g. in <c e g>~ g, the g will create a tie, but the c and e should
not trigger a warning!