Use callback to retrieve OctavateEight break-visibility from clef
Do not explicitly assign break-visibility to the OctavateEight,
use a callback to retrieve from clef. For one, this allows the user
to override the clef's break-visibility and affect the OctavateEight
without an explicit setting for it. On the other hand, by overriding
the break-visibility from the callback to a custom value (e.g.
all-invisible), this makes it finally possible to hide 8 from
a treble_8 clef for a tenor piece altogether, while keeping
the correct pitch (for MIDI etc.).